Move and divide utils
