+++ python/packages/ai/teams/ai/prompts/function_response_message.py
@@ -10,8 +10,8 @@ from typing import Any, List
 from botbuilder.core import TurnContext
 
 from ...state import MemoryBase
+from ...utils.to_string import to_string
 from ..tokenizers import Tokenizer
-from ..utilities import to_string
 from .message import Message
 from .prompt_functions import PromptFunctions
 from .rendered_prompt_section import RenderedPromptSection
