+++ python/packages/ai/teams/ai/prompts/sections/conversation_history_section.py
@@ -11,8 +11,8 @@ from typing import List
 from botbuilder.core import TurnContext
 
 from ....state import MemoryBase
+from ....utils.to_string import to_string
 from ...tokenizers import Tokenizer
-from ...utilities import to_string
 from ..message import Message
 from ..prompt_functions import PromptFunctions
 from ..rendered_prompt_section import RenderedPromptSection
