+++ python/packages/ai/tests/utils/__init__.py
@@ -5,3 +5,5 @@ Licensed under the MIT License.
 
 from .activity import *
 from .adapter import *
+from .test_snippet import *
+from .test_to_string import *
